- Updated DataSourceBag to use enum value.

+++ Rock.Blocks/CheckIn/Configuration/LabelDesigner.cs
@@ -66,7 +66,7 @@ namespace Rock.Blocks.CheckIn.Configuration
                 Name = dataSource.Name,
                 TextSubType = dataSource.TextSubType,
                 Category = dataSource.Category,
-                SupportedComparisionTypes = ( int ) dataSource.SupportedComparisionTypes,
+                SupportedComparisionTypes = dataSource.SupportedComparisionTypes,
                 CustomFields = dataSource.CustomFields,
                 FormatterOptions = dataSource.Formatter?.Options ?? new List<DataFormatterOptionBag>()
             };
